Refactor: Entity 추가에 따른 문제 관련 API 수정
MatchingService의 API들 증 Question 관련, usercode랑 Mastercode FK 참조하게 수정

Fixes: S09P12A107-135

+++ backend/src/main/java/com/ssafy/solive/api/service/MatchingServiceImpl.java
@@ -5,10 +5,14 @@ import com.ssafy.solive.api.request.QuestionModifyPutReq;
 import com.ssafy.solive.api.request.QuestionRegistPostReq;
 import com.ssafy.solive.common.exception.ImageUploadFailException;
 import com.ssafy.solive.common.exception.NoImageException;
+import com.ssafy.solive.db.entity.MasterCode;
 import com.ssafy.solive.db.entity.Question;
 import com.ssafy.solive.db.entity.QuestionPicture;
+import com.ssafy.solive.db.entity.User;
+import com.ssafy.solive.db.repository.MasterCodeRepository;
 import com.ssafy.solive.db.repository.QuestionPictureRepository;
 import com.ssafy.solive.db.repository.QuestionRepository;
+import com.ssafy.solive.db.repository.UserRepository;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
@@ -24,12 +28,17 @@ public class MatchingServiceImpl implements MatchingService {
 
     QuestionRepository questionRepository;
     QuestionPictureRepository questionPictureRepository;
+    UserRepository userRepository;
+    MasterCodeRepository masterCodeRepository;
 
     @Autowired
     public MatchingServiceImpl(QuestionRepository questionRepository,
-        QuestionPictureRepository questionPictureRepository) {
+        QuestionPictureRepository questionPictureRepository, UserRepository userRepository,
+        MasterCodeRepository masterCodeRepository) {
         this.questionRepository = questionRepository;
         this.questionPictureRepository = questionPictureRepository;
+        this.userRepository = userRepository;
+        this.masterCodeRepository = masterCodeRepository;
     }
 
     @Override
@@ -38,11 +47,26 @@ public class MatchingServiceImpl implements MatchingService {
         if (files.get(0).getSize() == 0) {
             throw new NoImageException();
         }
-        // TODO: Studnt Entity 구현 후, Req의 StudentId 에 대해 Illegal Arg 확인 필요
-        Question question = questionRepository.save(registInfo.toQuestion());
+        // TODO: Studnt Entity 구현 후, Student 지위 체크 과정 필요한 지 확인
+        User user = userRepository.findById(registInfo.getStudentId())
+            .orElseThrow(IllegalArgumentException::new);
+        MasterCode masterCode = masterCodeRepository.findById(registInfo.getMasterCodeId())
+            .orElseThrow(IllegalArgumentException::new);
+        String title = registInfo.getTitle();
+        String description = registInfo.getDescription();
+
+        Question question = Question.builder()
+            .user(user)
+            .masterCode(masterCode)
+            .title(title)
+            .description(description)
+            .build();
+
+        questionRepository.save(question);
         // TODO: JpaRepository의 save에서 Exception이 발생할 경우 어떻게 처리할 지 조사할 필요
         // TODO: 파일 업로드 관련 추가한 컬럼들 차후 상황에 따라 재고 필요
         String uploadFilePath = "C:/solive/image/";
+
         for (MultipartFile file : files) {
             String prefix = file.getOriginalFilename()
                 .substring(file.getOriginalFilename().lastIndexOf(".") + 1,
@@ -76,21 +100,30 @@ public class MatchingServiceImpl implements MatchingService {
 
     @Override
     public boolean deleteQuestion(QuestionDeletePutReq deleteInfo) {
+        // TODO: IllegalArg Exception 을 적절한 Custom Exception으로 대체
         Question question = questionRepository.findById(deleteInfo.getQuestionId())
             .orElseThrow(IllegalArgumentException::new);
-        // TODO: IllegalArg Exception 을 적절한 Custom Exception으로 대체
-        // TODO: Student Entity 만들어지면 인가 과정 작성 필요 -> 실패시 return false
-        question.deleteQuestion();
-        return true;
+        if (question.getUser().getId().equals(deleteInfo.getStudentId())) {
+            question.deleteQuestion();
+            return true;
+        }
+        return false;
     }
 
     @Override
     public boolean modifyQuestion(QuestionModifyPutReq modifyInfo) {
+        // TODO: IllegalArg Exception 을 적절한 Custom Exception으로 대체
         Question question = questionRepository.findById(modifyInfo.getQuestionId())
             .orElseThrow(IllegalArgumentException::new);
-        // TODO: IllegalArg Exception 을 적절한 Custom Exception으로 대체
-        // TODO: Student Entity 구현 완료 후 수정이 가능한 사용자인지 인가 과정 필요 -> 실패시 return false
-        question.modifyQuestion(question);
-        return true;
+        MasterCode masterCode = masterCodeRepository.findById(modifyInfo.getMasterCodeId())
+            .orElseThrow(IllegalArgumentException::new);
+        String title = modifyInfo.getTitle();
+        String description = modifyInfo.getDescription();
+
+        if (question.getUser().getId().equals(modifyInfo.getStudentId())) {
+            question.modifyQuestion(masterCode, title, description);
+            return true;
+        }
+        return false;
     }
 }
